chart_analysis_agent.py
import cv2
import numpy as np
from PIL import Image
from typing import Dict, Any, List
from .base_agent import BaseAgent, AgentResponse

class ChartAnalysisAgent(BaseAgent):
    """Agent responsible for visual analysis of financial charts"""
    
    def __init__(self):
        super().__init__(
            name="ChartAnalysisAgent",
            description="Analyzes financial charts for visual patterns and key elements"
        )
        
    async def process(self, input_data: Dict[str, Any]) -> AgentResponse:
        try:
            # Extract image data
            image_data = input_data.get('image')
            if not image_data:
                return AgentResponse(
                    success=False,
                    data={},
                    error="No image data provided"
                )
            
            # Convert image data to numpy array
            image = self._preprocess_image(image_data)
            
            # Perform visual analysis
            analysis_results = await self._analyze_chart(image)
            
            return AgentResponse(
                success=True,
                data=analysis_results
            )
            
        except Exception as e:
            return AgentResponse(
                success=False,
                data={},
                error=str(e)
            )
    
    def _preprocess_image(self, image_data: Any) -> np.ndarray:
        """Preprocess the input image for analysis"""
        if isinstance(image_data, str):  # If image path is provided
            image = cv2.imread(image_data)
        elif isinstance(image_data, np.ndarray):
            image = image_data
        else:
            # Convert PIL Image to numpy array
            image = np.array(image_data)
        
        # Convert to grayscale for better pattern recognition
        if len(image.shape) == 3:
            image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
        
        return image
    
    async def _analyze_chart(self, image: np.ndarray) -> Dict[str, Any]:
        """Analyze the chart for patterns and key elements"""
        results = {
            'patterns': self._detect_patterns(image),
            'key_levels': self._detect_key_levels(image),
            'trend_lines': self._detect_trend_lines(image),
            'volume_profile': self._analyze_volume_profile(image)
        }
        return results
    
    def _detect_patterns(self, image: np.ndarray) -> List[Dict[str, Any]]:
        """Detect chart patterns like head and shoulders, double tops/bottoms"""
        # TODO: Implement pattern detection using computer vision techniques
        return []
    
    def _detect_key_levels(self, image: np.ndarray) -> List[Dict[str, Any]]:
        """Detect support and resistance levels"""
        # TODO: Implement key level detection
        return []
    
    def _detect_trend_lines(self, image: np.ndarray) -> List[Dict[str, Any]]:
        """Detect trend lines in the chart"""
        # TODO: Implement trend line detection
        return []
    
    def _analyze_volume_profile(self, image: np.ndarray) -> Dict[str, Any]:
        """Analyze volume profile of the chart"""
        # TODO: Implement volume profile analysis
        return {} 