requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
opencv-python>=4.5.0
torch>=1.9.0
transformers>=4.11.0
pillow>=8.3.0
python-dotenv>=0.19.0
langchain>=0.0.150
openai>=0.27.0
pydantic>=1.8.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
ta>=0.10.0  # Technical Analysis library
yfinance>=0.1.63  # Yahoo Finance API 